+++ packages/libredwg-converter/src/AcDbEntitiyConverter.ts
@@ -188,10 +188,11 @@ export class AcDbEntityConverter {
         spline.controlPoints,
         spline.knots,
         spline.weights,
+        spline.degree,
         !!(spline.flag & 0x01)
       )
     } else if (spline.numberOfFitPoints > 0) {
-      return new AcDbSpline(spline.fitPoints, 'Uniform', !!(spline.flag & 0x01))
+      return new AcDbSpline(spline.fitPoints, 'Uniform', spline.degree, !!(spline.flag & 0x01))
     }
     return null
   }